+++ src/data/collections/array/d1/uninit/impl_traits.rs
@@ -9,7 +9,7 @@ use crate::{
     Ordering, Storage, _core::fmt,
 };
 
-impl<T, const CAP: usize, S: Storage> Drop for UninitArray<T, CAP, S> {
+impl<T, const CAP: usize, S: Storage> Drop for ArrayUninit<T, CAP, S> {
     fn drop(&mut self) {
         // SAFETY: we are only dropping initialized elements.
         for i in 0..self.len() {
@@ -21,7 +21,7 @@ impl<T, const CAP: usize, S: Storage> Drop for UninitArray<T, CAP, S> {
 }
 
 // AsRef
-impl<T, const CAP: usize, S: Storage> AsRef<[T]> for UninitArray<T, CAP, S> {
+impl<T, const CAP: usize, S: Storage> AsRef<[T]> for ArrayUninit<T, CAP, S> {
     /// Returns a shared slice of the initialized elements.
     fn as_ref(&self) -> &[T] {
         // SAFETY: the slice is constructed from the initialized range
@@ -29,7 +29,7 @@ impl<T, const CAP: usize, S: Storage> AsRef<[T]> for UninitArray<T, CAP, S> {
     }
 }
 // AsMut
-impl<T, const CAP: usize, S: Storage> AsMut<[T]> for UninitArray<T, CAP, S> {
+impl<T, const CAP: usize, S: Storage> AsMut<[T]> for ArrayUninit<T, CAP, S> {
     /// Returns an exclusive slice of the initialized elements.
     fn as_mut(&mut self) -> &mut [T] {
         // SAFETY: the slice  from the initialized range
@@ -37,7 +37,7 @@ impl<T, const CAP: usize, S: Storage> AsMut<[T]> for UninitArray<T, CAP, S> {
     }
 }
 // Borrow
-impl<T, const CAP: usize, S: Storage> Borrow<[T]> for UninitArray<T, CAP, S> {
+impl<T, const CAP: usize, S: Storage> Borrow<[T]> for ArrayUninit<T, CAP, S> {
     /// Returns a shared slice of the initialized elements.
     fn borrow(&self) -> &[T] {
         // SAFETY: we're only creating a slice from the initialized part
@@ -45,7 +45,7 @@ impl<T, const CAP: usize, S: Storage> Borrow<[T]> for UninitArray<T, CAP, S> {
     }
 }
 // AsMut
-impl<T, const CAP: usize, S: Storage> BorrowMut<[T]> for UninitArray<T, CAP, S> {
+impl<T, const CAP: usize, S: Storage> BorrowMut<[T]> for ArrayUninit<T, CAP, S> {
     /// Returns an exclusive slice of the initialized elements.
     fn borrow_mut(&mut self) -> &mut [T] {
         // SAFETY: we're only creating a slice from the initialized part
@@ -53,7 +53,7 @@ impl<T, const CAP: usize, S: Storage> BorrowMut<[T]> for UninitArray<T, CAP, S> 
     }
 }
 // T:Clone
-impl<T: Clone, const CAP: usize, S: Storage> Clone for UninitArray<T, CAP, S>
+impl<T: Clone, const CAP: usize, S: Storage> Clone for ArrayUninit<T, CAP, S>
 where
     S::Stored<[MaybeUninit<T>; CAP]>: Clone,
 {